% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_LIMSAPP.R
\name{check_and_set_commandArgs}
\alias{check_and_set_commandArgs}
\title{Manueel zetten van de commando argumenten}
\usage{
check_and_set_commandArgs(argvalues = NULL, n_min = 3)
}
\arguments{
\item{argvalues}{vector met de commando-argumenten (allemaal tussen quotes) zoals je zou verwachten als je het script laat runnen via de command line. Het 6e argument is doorgaans het eerste relevante argument en bevat normaliter de naam van het script, de andere argumenten kunnen variëren, maar vaak zijn dsn, uid en pwd het 7e, 8e en 9e argument, daarna gevolgd door scriptspecifieke parameters}

\item{n_min}{minimum aantal argumenten die nodig zijn voor gekozen wordt om argvalues in te vullen zoals in argvalues gespecifieerd.}
}
\value{
een lijst met de echte commandArgs() als die bestaan (meer dan n_min argumenten in commandArgs) of de zelf gecreëerde indien de commandArgs() slechts n_min argumenten bevat
}
\description{
Manueel zetten van de commando argumenten
}
\examples{
check_and_set_commandArgs(argvalues = c("","","","","","", par1 = "foobar", aantal = 5), 
                          n_min = 3)
}
