% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_parsing.R
\name{date_as_text}
\alias{date_as_text}
\title{Converteer datumvector of POSIXct vector naar text in formaat yyyy-mm-dd}
\usage{
date_as_text(x, na_result = "NODATE")
}
\arguments{
\item{x}{vector met datums (character)}

\item{na_result}{resultaat indien geen geldige datum}
}
\value{
character vector with datums
}
\description{
Converteer datumvector of POSIXct vector naar text in formaat yyyy-mm-dd
}
