% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ddPCR.R
\name{ddpcr_calculate_fit}
\alias{ddpcr_calculate_fit}
\title{Berekeningen van een trendlijn in de qc grafiek voor digital droplet PCR}
\usage{
ddpcr_calculate_fit(
  x,
  y,
  start = c(a = 1, b = 1, c = 1),
  estim_formula = "connect",
  loess_span = 0.5
)
}
\arguments{
\item{x}{vector met de x-coordinaten in de grafiek}

\item{y}{vector met de y-coordinaten in de grafiek}

\item{start}{startwaarden voor het berekenen van de verschillende formules.
Vector met 3 numerieke elementen.}

\item{estim_formula}{welke formule gebruikt wordt. Keuze tussen connect,
growth, loess, polynomial, elu en sigmoid. Standaard wordt connect gebruikt,
welke gewoon de gemiddelde respons berekend bij elke x waarde
en deze als een lijn verbindt, wat het voordeel heeft
dat er geen aannames moeten gemaakt worden,
maar als nadeel dat de curve er niet altijd mooi uitziet}

\item{loess_span}{waarde die specifiek aan de loess functie wordt doorgegeven
als f (smoother span)}
}
\value{
een dataset met een x-waarde en een y-waarde.
De x is ofwel de echte x-waarden (voor bv connect),
maar gebruikt de range van x opgedeeld in 100 stukjes
voor een exponentiële functie
}
\description{
Berekeningen van een trendlijn in de qc grafiek voor digital droplet PCR
}
