% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_report_fields.R
\name{get_all_report_fields}
\alias{get_all_report_fields}
\title{Show available fields}
\usage{
get_all_report_fields()
}
\value{
dataset containing 2 colums: the name of a field in the database
and the corresponding description
}
\description{
Show available fields
}
\examples{
{
get_all_report_fields()
}
}
