% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_config_info.R, R/lims_report.R
\name{get_report_config_info}
\alias{get_report_config_info}
\title{Get report config information}
\usage{
get_report_config_info(template = "default")

get_report_config_info(template = "default")
}
\arguments{
\item{template}{a character vector of templates you want info from.
If "all" then all templates are shown}
}
\value{
dataset wih report template information

dataset wih report template information
}
\description{
Get report config information

Get report config information
}
\examples{
{
get_report_config_info(template = "default")
}
{
get_report_config_info(template = "default")
}
}
