% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ddPCR.R
\name{ggplot.ddPCR}
\alias{ggplot.ddPCR}
\title{GGplot of ddPCR data (generated by ddPCR_qcalc)}
\usage{
\method{ggplot}{ddPCR}(data, mapping = NULL, colors = c("black", "red",
  "blue", "green4"), shapes = c(-1, 1, 4), linetypes = c("solid",
  "blank", "solid", "dashed"), LOD_label = "LOD", LOQ_label = "LOQ",
  xlab = expression(paste(log[10], "(calculated DNA concentration)")),
  ylab = expression(paste(log[10], "(measured DNA concentration)")),
  log_add = 1e-04)
}
\arguments{
\item{data}{dataset generated from ddPCR_qccalc. Must have at least these columns:
\itemize{
  \item log10Concentration
  \item log10Theoretical
  \item which_data groeperingsvariabele bestaande uit 3 niveau's. De naam is vrij te kiezen, maar de alfabetische volgorde zal bepalen welke kleur en lijntype eraan gematcht worden
  \item log10LOQ
  \item log10LOD
}}

\item{mapping}{wordt niet gebruikt in dit geval, maar is nodig om de ggplot functie te kunnen gebruiken}

\item{colors}{vector van lengte 4 met de kleuren van de verschillende plotdelen die bepaald worden door de volgorde van de niveau's in which_data
\itemize{
  \item kleur 1 is het kleur voor de eerste groep (standaard de gefitte lijn)
  \item kleur 2 is het kleur voor de tweede groep (standaard de observaties)
  \item kleur 3 is het kleur voor de derde groep (standaard de theoretische concentratie)
  \item kleur 4 is het kleur voor de LOD en LOQ lijnen
}}

\item{shapes}{numerieke vector van lengte 3 die de symbolen overeenkomstig de groepen in which_data definieert. Bij een negatief getal wordt het punt niet getoond, symbool met nummer 1 is een lege bol, symbool met nummer 4 is een kruisje. Standaard wordt de eerste shape negatief gezet, omdat dit standaard overeenkomt met de gefitte lijn, waarvoor we de punten niet willen zien}

\item{linetypes}{vector in dezelfde volgorde als colors die de lijntypes definieert. Voor een lijn die niet getekend moet worden geef je de waarde "blank", een volle lijn is "solid", een stippelliljn "dashed" en dan heb je nog heel wat andere types die je kan terugvinden via google "ggplot linetypes"}

\item{LOD_label}{tekst die bij de LOD lijn moet gezet worden}

\item{LOQ_label}{tekst die bij de LOQ lijn moet gezet worden}

\item{xlab}{astitel voor de x-as}

\item{ylab}{astitel voor de y-as}

\item{log_add}{hoeveel wordt bij de concentratie geteld om de log te nemen (log10(variabele + log_add)). Deze waarde wordt gebruikt om enkel de datarijen te selecteren met een concentratie groter dan 0, en moet overeenstemmen met de log_add in ddPCR_qc_calc}
}
\value{
ggplot object met 5 lagen: de geobserveerde waarden, de fit door de niet-nulwaarden van de geobserveerde waarden, de theoretische waarden, een vertikale lijn voor LOD, een vertikale lijn voor LOQ
}
\description{
GGplot of ddPCR data (generated by ddPCR_qcalc)
}
