% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbolims-package.R
\docType{package}
\name{inbolims-package}
\alias{inbolims}
\alias{inbolims-package}
\title{inbolims: Utilities to query the INBO lab analysis datawarehouse}
\description{
Some utilities that expand the Lims program in INBO. It mainly focuses on post-processing in connection with Quality Assurance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/inbolims}
  \item \url{https://inbo.github.io/inbolims/}
  \item Report bugs at \url{https://github.com/inbo/inbolims/issues}
}

}
\author{
\strong{Maintainer}: Pieter Verschelde \email{pieter.verschelde@inbo.be} (\href{https://orcid.org/0000-0002-9199-421X}{ORCID}) (Research Institute for Nature and Forest (INBO))

}
\keyword{internal}
