% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_inbo_domain.R
\name{is_on_inbo_domain}
\alias{is_on_inbo_domain}
\title{Check if on inbo domain}
\usage{
is_on_inbo_domain(domain = "inbo.be")
}
\arguments{
\item{domain}{character string containing the dns domain}
}
\value{
logical TRUE if on domain, FALSE else
}
\description{
Check if on inbo domain
}
