% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_LIMSDB.R
\name{lims_db_connect}
\alias{lims_db_connect}
\title{Title}
\usage{
lims_db_connect(server = "inbo-sql07-prd.inbo.be",
  database = "D0015_00_Lims", uid, pwd)
}
\arguments{
\item{server}{name of the LIMS server}

\item{database}{name of the LIMS database}

\item{uid}{application username (that as writing rights on the db, so this is not you username)}

\item{pwd}{applciation username password}
}
\value{
db connection
}
\description{
Title
}
\examples{
\dontrun{
lims_db_connect(uid = "me", pwd = "123456") #should not work
}
}
