% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_plots.R
\name{lims_density}
\alias{lims_density}
\title{Densityplot van het lims resultaat}
\usage{
lims_density(data, group = NULL, xlab, ylab, file = NA, show = TRUE,
  suffix = "", dpi = 300, width = 7, height = 5, ncol = 1,
  quantiles = TRUE, ...)
}
\arguments{
\item{data}{lims dataset to plot}

\item{group}{charactervariabele die de groeperingskolom bepaalt}

\item{xlab}{naam x-as}

\item{ylab}{naam y-as}

\item{file}{bestandsnaam waar de figuur wordt bewaard (in de werkdirectory)}

\item{show}{toon de figuur op het scherm}

\item{suffix}{korte naam voor de figuur, deze zal scatter_date_suffix.png noemen}

\item{dpi}{resolutie}

\item{width}{breedte in inch}

\item{height}{hoogte in inch}

\item{ncol}{Als een groepvariabele gedefinieerd is, in hoeveel kolommen moet de figuren getoond worden}

\item{quantiles}{indien TRUE worden de 2.5, 25, 50, 75 en 97.5 percentielen als lijnen weergegeven}

\item{...}{parameters die aan geom_density kunnen meegegeven worden}
}
\value{
gg densityplot
}
\description{
Densityplot van het lims resultaat
}
