% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_LIMSDWH.R
\name{lims_get_results}
\alias{lims_get_results}
\title{Lees ruwe data in}
\usage{
lims_get_results(conn = NULL, sampletype = c("NULL", "DUP",
  "SUBSAMPLE"), contract = NULL, project = NULL, analysis = NULL,
  component = NULL, matrix = NULL, start = end - 4 * 31536000,
  end = Sys.time(), samplestatus = c("C", "A"), resultstatus = c("E",
  "M", "A"), n_max = 1e+05, print_query = TRUE)
}
\arguments{
\item{conn}{connectie met de databank bekomen via lims_connect(). Indien NULL zal de standaard lims_connect() door deze functie worden uitgevoerd, alsook een disconnect}

\item{sampletype}{vector met gekozen sampletypes. NULL geeft enkel de hoofdstalen. Standaard wordt NULL, DUP en SUBSAMPLE gekozen, omdat dit de 3 types zijn die de gewone stalen weergeven. PBL en BLANK zijn de blanco's}

\item{contract}{numerieke vector met gekozen contractnummers. NULL geeft ze allemaal}

\item{project}{charactervector met gekozen projecten. NULL geeft ze allemaal}

\item{analysis}{charactervector met gekozen analyses. NULL is allemaal}

\item{component}{charactervector met gekozen componenten. NULL is allemaal}

\item{matrix}{charactervector met gekozen matrices. NULL is allemaal}

\item{start}{datumtijd variabele met de starttijd}

\item{end}{datumtijd variabele met de einddtijd}

\item{samplestatus}{welke sample status (standaard C en A). NULL geeft ze allemaal}

\item{resultstatus}{welke resultaatstatus (standaard E en M en A). NULL geeft ze allemaal}

\item{n_max}{hoeveel resultaten maximaal tonen}

\item{print_query}{toon de querystring op het scherm}
}
\value{
dataset met de gevraagde resultaten
}
\description{
Lees ruwe data in
}
\examples{
\dontrun{
data <- lims_get_results(lims_connect(),project = "I-18W001-01")
data <- lims_get_results(analysis = "LOI_AS_OVEN_550", start = "2018-12-30")
data <- lims_get_results(analysis = "CACO3_TIT_V", component = c("CaCO3\%", "Gewicht"), 
                         start = "2018-01-01", end = "2019-01-01")
}
}
