% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_LIMSDWH.R
\name{lims_mogelijkheden}
\alias{lims_mogelijkheden}
\title{Toon lijst met mogelijke keuzes voor de tabelvelden}
\usage{
lims_mogelijkheden(conn = NULL, keuze = c("LimsAnalysisName", "Matrix",
  "SampleType", "Project", "StatusCode"))
}
\arguments{
\item{conn}{connectie met de databank. Indien NULL wordt de standaard lims_connect() gebruikt}

\item{keuze}{keuze waarvan je een lijst van mogelijkheden wilt. Je kan hier kiezen tussen LimsAnalyseNaam, Matrix, SampleType, Project, StatusCode}
}
\value{
data.frame met alle waarden van de keuzelijsten, zoals in parameter keuze gedefinieerd
}
\description{
Toon lijst met mogelijke keuzes voor de tabelvelden
}
\examples{
\dontrun{
lims_mogelijkheden(keuze = "Matrix")
}
}
