% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_quantile.R
\name{lims_quantile}
\alias{lims_quantile}
\title{Kwantielberekening}
\usage{
lims_quantile(data, group = NULL, probs = c(0, 0.025, 0.25, 0.5, 0.75,
  0.975, 1), file = NA, ...)
}
\arguments{
\item{data}{lims data waarvoor de kwantielen berekend moeten worden}

\item{group}{geef hier de kolomnaam in die dient om de groepen te selecteren waarvoor de kwantielen berekend worden}

\item{probs}{de probabiliteiten waarvoor de kwantielen moeten berekend worden. Een vector van elementen tussen 0 en 1}

\item{file}{nog niet gebruikt. Het bestand waarnaar de tabel geschreven wordt}

\item{...}{extra parameters die aan de kwanitelfunctie kunnen meegegeven worden}
}
\value{
tibbble met de waarden horende bij ieder kwantiel. Indien groepen gekozen, een aparte kolom per groep
}
\description{
Kwantielberekening
}
