% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R
\name{lims_report_data}
\alias{lims_report_data}
\title{Haal rapportdata uit LIMS DWH}
\usage{
lims_report_data(
  project,
  template = "default",
  deployment = "prd",
  show_query = FALSE
)
}
\arguments{
\item{project}{charactervector met projectnamen}

\item{template}{indien "default" wordt de standaardquery uitegevoerd, indien iets anders moet  dit een geldige query zijn voor het DWH en minstens de outputvelden ORIGINAL_SAMPLE, ANALYSIS, COMPONENT en ENTRY bevatten}

\item{deployment}{default "prd", "uat" indien op de ontwikkelomgeving gewerkt wordt}

\item{show_query}{indien TRUE toon de query net voordat deze uitgevoerd wordt}
}
\value{
data.frame met minstens de velden ORIGINAL_SAMPLE, ANALYSIS, COMPONENT en ENTRY
}
\description{
Haal rapportdata uit LIMS DWH
}
\examples{
\dontrun{
reportdata <- lims_report_data(project = c("I-19W001-01"))
}
}
