% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R
\name{lims_report_samples}
\alias{lims_report_samples}
\title{Verkrijg de sample metadata}
\usage{
lims_report_samples(reportdata)
}
\arguments{
\item{reportdata}{data verkregen uit de functie lims_report_data}
}
\value{
dataset met sample informatie
}
\description{
Verkrijg de sample metadata
}
\examples{
\dontrun{
reportdata <- lims_report_data(project = c("I-19W001-01"))
sampledata <- lims_report_samples(reportdata)
}

}
