% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R
\name{lims_report_sql_create}
\alias{lims_report_sql_create}
\title{Vul de query dynamisch in}
\usage{
lims_report_sql_create(project = NULL, deployment = "prd")
}
\arguments{
\item{project}{charactervector met projectnamen}

\item{deployment}{"prd" indien op productie, "uat" indien in developmentomgeving}
}
\value{
lange string met de query
}
\description{
Vul de query dynamisch in
}
