% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R
\name{lims_report_xtab}
\alias{lims_report_xtab}
\title{Maak kruistabel van de ingelezen rapportdata}
\usage{
lims_report_xtab(reportdata)
}
\arguments{
\item{reportdata}{data verkregen uit de functie lims_report_data}
}
\value{
kruistabel met resultaten
}
\description{
Maak kruistabel van de ingelezen rapportdata
}
