% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_sample_information.R
\name{lims_sample_information}
\alias{lims_sample_information}
\title{Verkrijg de sample informatie}
\usage{
lims_sample_information(con, project, fields = "template:default")
}
\arguments{
\item{con}{connectie naar lims dwh}

\item{project}{een character string met projecten,
eventueel met SQL jokertekens "\%" en "_"}

\item{fields}{welke velden wil je terugkrijgen, #nolint
indien het begint met "template:" wordt een staaltemplate gekozen,
anders is dat gewoon een vector van velden. Je kan "template:"
en eigen veldnamen combineren bv.
c("template:default", "FieldObserver", "FieldSamplingDate")}
}
\value{
tibble met sample informatie
}
\description{
Verkrijg de informatie van stalen voor gekozen projecten
}
\examples{
library(inbolims)
con <- lims_connect()
sample_info <- lims_sample_information(con, project = "I-20G\%")
sample_info <-
 lims_sample_information(
   con,
   project = c("I-19G024-01", "I-20G\%"),
   fields = c("template:default", "FieldSamplingDate"))
}
