% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_sample_overview.R
\name{lims_sample_overview}
\alias{lims_sample_overview}
\title{Lims overzicht projectstalen}
\usage{
lims_sample_overview(
  conn,
  customer = NULL,
  project = NULL,
  lab = c("VAST", "WATER", "GENETICA"),
  status = c("I", "P", "C", "A"),
  sql_filter = NULL,
  sql_order = NULL,
  start = NULL,
  end = NULL,
  show_query = FALSE,
  show_fields = c("")
)
}
\arguments{
\item{conn}{datawarehouse connection, typically the result of limsdwh_connect()}

\item{customer}{vector of selected customers. If customer is specified with !, then this will result in a not selection of the customer. e.g customer = c("GEN_DIV", "GEN_BOS") or customer = c("!TESTKLANT")}

\item{project}{vector of selected projects. "\%" and "\_" can be used as wildcharts as in SQL}

\item{lab}{vector met de gekozen labo's: mogelijkheden zijn VAST, WATER, GENETICA}

\item{status}{vector die bepaalt welke statussen van stalen die getoond worden. I (incompleet), P (bezig), C (compleet), A (geauthoriseerd), X (gecanceld) zijn de mogelijkheden}

\item{sql_filter}{toevoeging aan het where statement in de gegenereerde SQL code}

\item{sql_order}{toevoeging van een custom ordering via SQL code}

\item{start}{vroegste logindatum van het staal}

\item{end}{laatste logindatum van het staal}

\item{show_query}{toon de gegenereerde SQL query}

\item{show_fields}{niet gebruikt}
}
\value{
data.frame with selected samples
}
\description{
Lims overzicht projectstalen
}
\examples{
\dontrun{
conn <-limsdwh_connect()
data <- lims_sample_overview(conn)
}
}
