% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_plots.R
\name{lims_scatter_nr}
\alias{lims_scatter_nr}
\title{Scatterplot van het resultaat van de laatste punten}
\usage{
lims_scatter_nr(
  data,
  group = NULL,
  xlab,
  ylab,
  clab,
  file = NA,
  show = TRUE,
  suffix = "",
  dpi = 300,
  width = 7,
  height = 5,
  add_path = TRUE,
  ...
)
}
\arguments{
\item{data}{lims dataset to plot}

\item{group}{charactervariabele die de groeperingskolom bepaalt}

\item{xlab}{naam x-as}

\item{ylab}{naam y-as}

\item{clab}{naam van de kleur-as}

\item{file}{bestandsnaam waar de figuur wordt bewaard (in de werkdirectory)}

\item{show}{toon de figuur op het scherm}

\item{suffix}{korte naam voor de figuur, deze zal scatter_date_suffix.png noemen}

\item{dpi}{resolutie}

\item{width}{breedte in inch}

\item{height}{hoogte in inch}

\item{add_path}{verbind de punten met een lijn}

\item{...}{parameters die aan opgeroepen subfuncties kunnen meegegeven worden}
}
\value{
gg scatterplot
}
\description{
Scatterplot van het resultaat van de laatste punten
}
