% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_table_fields.R
\name{lims_table_fields}
\alias{lims_table_fields}
\title{Get all availaible column names from a table in the data warehouse}
\usage{
lims_table_fields(con, table_name, export_type = "character")
}
\arguments{
\item{con}{connection to the lims Dwh. Typically by lims_connect()}

\item{table_name}{name of the table from which the fields are wanted.
Important tables are dimSample and factResult.}

\item{export_type}{how to return: "character", "tibble" or "data.frame"}
}
\value{
character vector with column names of the table
}
\description{
Get all availaible column names from a table in the data warehouse
}
\examples{
\dontrun{
library(inbolims)
con <- lims_connect()
lims_table_fields(con, "dimSample")
}
}
