% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_test_information.R
\name{lims_test_information}
\alias{lims_test_information}
\title{Show test overview}
\usage{
lims_test_information(
  conn,
  project,
  include_non_reportable = TRUE,
  include_components = TRUE,
  include_suffix = FALSE
)
}
\arguments{
\item{conn}{database connection via lims_connect}

\item{project}{project code, e.g. "I-19W001-01",
sql joker characters are allowed, e.g. "I-19W%"}

\item{include_non_reportable}{logical, if TRUE include non reportable tests,}

\item{include_components}{logical, if TRUE include components in the output}

\item{include_suffix}{logical, if TRUE include suffix
(TestReplicateCount__ResultReplicate)}
}
\value{
data frame with test information
}
\description{
Show test overview
}
\examples{
\dontrun{
conn <- lims_connect()
lims_test_information(conn, project = "I-19W001-01")
lims_test_information(conn,
  project = "I-19W001-01",
  include_components = FALSE
)
lims_test_information(conn,
  project = "I-19W001-01",
  include_non_reportable = FALSE
)
}
}
