% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R
\name{lims_xtab_to_csv}
\alias{lims_xtab_to_csv}
\title{Kruistabel naar csv wegschtrijven met toevoeging header}
\usage{
lims_xtab_to_csv(data, path)
}
\arguments{
\item{data}{kruistabeldata uit lims_report_xtab}

\item{path}{pad waar de file geschreven moet worden}
}
\value{
csv file
}
\description{
Kruistabel naar csv wegschtrijven met toevoeging header
}
