% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limsdwh_connect.R
\name{limsdwh_connect}
\alias{limsdwh_connect}
\title{Connect to Data warehouse.}
\usage{
limsdwh_connect(deployment = "prd")
}
\arguments{
\item{deployment}{"prd" betekent uit de productieserver halen, "uat" is enkel voor testen met het datawarehouse op de uat server. De toegangsrechten moeten voor uat via referentiebeheer in windows gedefinieerd zijn}
}
\value{
database connection object
}
\description{
De connectie kan enkel gebeuren via windows authentificatie. Er wordt gebruik gemaakt van DBI::dbConnect en odbc::odbc().
}
\examples{
\dontrun{
conn <- limsdwh_connect()
sql <- "select top(10) * from factResult"
dfResults <- dbGetQuery(conn, sql)
str(dfResults)
}
}
