% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limsdwh_report.R
\name{limsdwh_report}
\alias{limsdwh_report}
\title{Haal resultaten uit het LIMS datawarehouse}
\usage{
limsdwh_report(
  conn,
  config = "DEFAULT",
  version = NULL,
  project = NULL,
  extra = NULL,
  ...
)
}
\arguments{
\item{conn}{connectiestring naar de database. Wordt aangemaakt via dwh_connect() of manueel}

\item{config}{character string of data.frame die aangeeft welke configuratie moet gebruikt worden (zie \link{limsdwh_report_query})}

\item{version}{indien NULL wordt de laatste versie van het opgegeven reporttype gebruikt, indien hier gespecifieerd wordt de gekozen versie gebruikt.}

\item{project}{naam van een project, bv "I-19W001-01", of meerdere namen als vector bv. c("I-19W001-01", "I-19W001-02"). Als derde mogelijkheid kan je ook 1 project selecteren met SQL joker karakters als \% en _. Als je alle projecten van contract 19W001 wil kan je dit als volgt specificeren: "\%19W001\%"}

\item{extra}{eventueel extra criteria toe te voegen aan de SQL, moet beginnen met 'and '}

\item{...}{other arguments, passed to limsdwh_report_convert_config_to_query like "show_query"}
}
\value{
data.frame uit het lims datawarehouse met alle analyseresultaten die gerapporteerd worden door het labo
}
\description{
Haal resultaten uit het LIMS datawarehouse
}
\examples{
\dontrun{
conn <- limsdwh_connect()
reportdata <- limsdwh_report(conn, project = 'I-17W003-01')
}

}
