% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limsdwh_report.R
\name{limsdwh_report_sample}
\alias{limsdwh_report_sample}
\title{Staalinformatie halen uit de ingelezen lims datawarehouse data}
\usage{
limsdwh_report_sample(df)
}
\arguments{
\item{df}{dataset ingelezen via limsdwh_report}
}
\value{
data.frame met enkel de staalinformatie
}
\description{
Staalinformatie halen uit de ingelezen lims datawarehouse data
}
