% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limsdwh_report.R
\name{limsdwh_report_xtab}
\alias{limsdwh_report_xtab}
\title{Kruistabel maken van de gegevens uit het lims datawarehouse}
\usage{
limsdwh_report_xtab(df, link = "Labocode", value = "WaardeRuw",
  collapse_repli = FALSE)
}
\arguments{
\item{df}{dataset die bekomen is via limsdwh_report}

\item{link}{naam waarop de stalen gelinkt worden in de kruistabel (slechts 1 rij per link), behalve als collapse_repli op FALSE staat, dan wordt sowieso de labocode gebruikt}

\item{value}{de kolomnaam uit df die de cellen zal bemannen in de kruistabel}

\item{collapse_repli}{Indien FALSE, dan wordt er gelinkt op de labocode ipv de link, indien "first" dan wordt de eerste meting genomen, indien "last" de laatste meting en indien "avg" het gemiddelde tussen de metingen}
}
\value{
dataset met als rijen de staalidentificatie en de kolommen alle resultaten van alle testen
}
\description{
Kruistabel maken van de gegevens uit het lims datawarehouse
}
