% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_parsing.R
\name{link_labo_id}
\alias{link_labo_id}
\title{Link het labo_id met het externe id}
\usage{
link_labo_id(
  conn,
  data,
  labo_id_col = "sample",
  extern_id_col = "FieldSampleID"
)
}
\arguments{
\item{conn}{db connectie}

\item{data}{dataset waarvan het labo_id gelinkt moet worden aan het extern id}

\item{labo_id_col}{naam van de kolom die het labo_id bevat}

\item{extern_id_col}{naam van de kolom waar je het externe id in wil}
}
\value{
dezelfde dataset met een extra kolom die het extern staalid bevat
}
\description{
Link het labo_id met het externe id
}
