% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R
\name{parse_sql_report_query}
\alias{parse_sql_report_query}
\title{Parse the query based on the template information and the chosen projects}
\usage{
parse_sql_report_query(template, project)
}
\arguments{
\item{template}{dataset containing the template_information.
The structure should be exact as the file shipped with the package}

\item{project}{character string of projects to filter}
}
\value{
SQL server query
}
\description{
Parse the query based on the template information and the chosen projects
}
