% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtp_parse_texture_content.R
\name{parse_texture_content}
\alias{parse_texture_content}
\title{Parste texture output file to R readable format}
\usage{
parse_texture_content(filename, delim = "\\t", verbose = TRUE)
}
\arguments{
\item{filename}{path to the output file to be parsed}

\item{delim}{delimiter of the file}

\item{verbose}{moet output getoond worden tijdens de uitvoering?}
}
\value{
parsed dataset with clear column names and data
}
\description{
Parste texture output file to R readable format
}
