% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lims_data.R
\name{read_lims_data}
\alias{read_lims_data}
\title{Haal rapportdata uit LIMS DWH}
\usage{
read_lims_data(
  connection,
  project,
  sql_template = "default",
  sample_types = c("Project"),
  include_non_reportable = FALSE,
  show_query = FALSE,
  custom_fields = NULL,
  custom_where_clause = NULL,
  custom_sql_query = NULL,
  deployment = "prd"
)
}
\arguments{
\item{connection}{DBI connection object (see odbc::dbConnect())}

\item{project}{charactervector met projectnamen}

\item{sql_template}{indien "default" wordt de standaardquery uitgevoerd,
indien "all" worden alle voorziene velden geïmporteerd,
bij "minimal" worden enkel de essentiële velden geïmporteerd}

\item{sample_types}{NOT USED FOR THE MOMENT charactervector
met de types staal die je wil inlezen
c("Project", "DUP", "BLANK", "CONTROL", "PRBLANCO", "STANDARD")
Voor genetica heb je nog andere stalen als QC_METHOD, SUBSAMPLE, ...}

\item{include_non_reportable}{Voorlopig is enkel "prd" actief ondersteund}

\item{show_query}{indien TRUE toon de query op het scherm
net voordat deze uitgevoerd wordt,
je kan deze eventueel kopiëren en aanpassen en doorgeven aan custom_sql_query}

\item{custom_fields}{charactervector die controleert
welke veldnamen je in de output ziet.
De velden ORIGINAL_SAMPLE, ANALYSIS, COMPONENT en ENTRY
worden altijd standaard meegeleverd.
Via de functie get_available_report_fields kan je alle namen
van de velden zien.}

\item{custom_where_clause}{custom SQL where clause dat geplakt wordt aan
"where PROJECT in ('project_x')",
dus een where clause dat je specificeerd als
" ResultNumeric > 0 OR ResultNumeric < 100,
zal vertaald worden naar de sql where clause:
"where PROJECT in ('project_x')
AND (ResultNumeric > 0 OR ResultNumeric < 100)"}

\item{custom_sql_query}{geldige Sql Query bruikbaar voor de connectie.
Indien dit veld actief is, wordt sql_template,
custom_fields en custom_where_clause genegeerd}

\item{deployment}{default "prd", "uat" indien op de ontwikkelomgeving
gewerkt wordt.}
}
\value{
data.frame met minstens de velden ORIGINAL_SAMPLE,
ANALYSIS, COMPONENT en ENTRY
}
\description{
Haal rapportdata uit LIMS DWH
}
\examples{
\dontrun{
conn <- lims_connect()
reportdata <- read_lims_data(conn,
  project = c("I-19W001-01"),
  sql_template = "default"
)
}
}
