% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fields_from_template.R
\name{sample_fields_from_template}
\alias{sample_fields_from_template}
\title{Velden uit sample_field_template}
\usage{
sample_fields_from_template(template)
}
\arguments{
\item{template}{sample_fields keuze}
}
\value{
character vector met kolomnamen
}
\description{
Velden uit sample_field_template
}
\examples{
sample_fields_from_template("default")
}
