% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_data_read_limsdata.R
\name{select_lims_object}
\alias{select_lims_object}
\title{Toon lijst met mogelijke keuzes voor de tabelvelden}
\usage{
select_lims_object(conn, keuze, how = c("list", "show"),
  multiple = FALSE, analyse = NULL, ...)
}
\arguments{
\item{conn}{connectie met de databank. Indien NULL wordt de standaard lims_connect() gebruikt}

\item{keuze}{keuze waarvan je een lijst van mogelijkheden wilt. Je kan hier kiezen tussen Analyse, Matrix, SampleType, Project, StatusCode, Component}

\item{how}{hoe wil je dat de informatie getoond wordt; Indien "show" toont dit gewoon de mogelijkheden, indien "list" kan je een of meerdere objecten selecteren uit een getoonde lijst}

\item{multiple}{if TRUE meerdere elementen kunnen geselecteerd worden, indien FALSE slechts 1 element.}

\item{analyse}{alleen gebruikt om componenten te selecteren. indien NULL krijg je een keuzelijst, anders wordt de hier gekozen waarde gebruikt}

\item{...}{andere argumenten die gebruikt kunnen worden in de functies \link{select.list} of \link{View}}
}
\value{
data.frame met alle waarden van de keuzelijsten, zoals in parameter keuze gedefinieerd, ofwel indien how = "list", dan krijg je een selectiemogelijkheid, en wordt de selectie teruggegeven
}
\description{
Toon lijst met mogelijke keuzes voor de tabelvelden
}
\examples{
\dontrun{
#zorg dat je een connectiebestand met de naam conn hebt
select_lims_object(conn, keuze = "Matrix", how = "list", multiple = TRUE)
}
}
