% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_available_templates.R
\name{show_available_templates}
\alias{show_available_templates}
\title{Toon beschikbare templates}
\usage{
show_available_templates()
}
\value{
data.frame with template name and number of fields
and a description of the template
}
\description{
Toon beschikbare templates
}
