% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ddPCR.R
\name{summary.ddPCR}
\alias{summary.ddPCR}
\title{Samenvattende QC statistieken voor de digital droplet PCR outputfile}
\usage{
\method{summary}{ddPCR}(object, ...)
}
\arguments{
\item{object}{data die gegenereerd wordt door de functie ddPCR_qc_calc}

\item{...}{voorlopig niet gebruikt}
}
\value{
data.frame met de samenvattende statistieken
}
\description{
Samenvattende QC statistieken voor de digital droplet PCR outputfile
}
