% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtp_tex_csv_2_json.R
\name{tex_csv_2_json}
\alias{tex_csv_2_json}
\title{Converteer textuurfiles naar JSON}
\usage{
tex_csv_2_json(fullfilename)
}
\arguments{
\item{fullfilename}{path of the file to be converted}
}
\value{
jsonfile
}
\description{
Converteer de bestanden die uit het textuurtoestel komen naar een json-file
}
