% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_data_write_db_credentials.R
\name{write_db_credentials}
\alias{write_db_credentials}
\title{Write db credentials}
\usage{
write_db_credentials(file = "dbcredentials.txt", dsn = NULL,
  uid = "User001", pwd = "123456")
}
\arguments{
\item{file}{bestandsnaam waar je de database credentials wil bewaren}

\item{dsn}{de naam van de gegevensbron (zie ODBC sources in windows)}

\item{uid}{de gebruikersnaam voor de db}

\item{pwd}{het passwoord voor de db voor de gebruikersnaam}
}
\value{
writes a text file with the database credentials
}
\description{
Write db credentials
}
\examples{
\dontrun{
write_db_credentials(file = "dbcredentials.txt", dsn = "MyDataSource", uid = "Me", pwd = "1234")
}
}
