% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_parsing.R
\name{write_texture_files}
\alias{write_texture_files}
\title{Schrijf textuurdata weg per staal in de target directory}
\usage{
write_texture_files(target_path, data, verbose = TRUE)
}
\arguments{
\item{target_path}{locatie waar de files moeten terechtkomen}

\item{data}{textuurdata die weggeschreven moet worden}

\item{verbose}{toon output terwijl de routine loopt}
}
\value{
er wordt een file weggeschreven per aanwezig staal
}
\description{
Schrijf textuurdata weg per staal in de target directory
}
