% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R
\name{lims_measured_parameters}
\alias{lims_measured_parameters}
\title{Title}
\usage{
lims_measured_parameters(data, plot = TRUE, log = TRUE)
}
\arguments{
\item{data}{lims report data (from lims_report_data)}

\item{plot}{logical if values are to be put in a graph}

\item{log}{when plot, use the log-scale?}
}
\value{
list with measured parameters and some base statistics
}
\description{
Title
}
\examples{
\dontrun{
view(lims_measured_parameters(reportdata))
}
}
