% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report_export.R
\name{lims_report_export}
\alias{lims_report_export}
\title{Kruistabel naar csv wegschtrijven met toevoeging header}
\usage{
lims_report_export(data, path)
}
\arguments{
\item{data}{kruistabeldata uit lims_report_xtab}

\item{path}{pad waar de file geschreven moet worden}
}
\value{
csv file
}
\description{
Kruistabel naar csv wegschtrijven met toevoeging header
}
