% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_report.R, R/lims_report_xtab.R
\name{lims_report_xtab}
\alias{lims_report_xtab}
\title{Maak kruistabel van de ingelezen rapportdata}
\usage{
lims_report_xtab(
  reportdata,
  resulttype = "measured",
  sample_fields = c("Project", "ExternSampleID")
)

lims_report_xtab(
  reportdata,
  resulttype = "measured",
  sample_fields = c("Project", "ExternSampleID")
)
}
\arguments{
\item{reportdata}{data verkregen uit de functie lims_report_data}

\item{resulttype}{"formatted" voor de geformatteerde waarde bv <1.20
of "measured" de originele waarde bv 1.185455}

\item{sample_fields}{Vector of sample fields that should appear
in the xtab file which are present in the report data}
}
\value{
kruistabel met resultaten

kruistabel met resultaten
}
\description{
Maak kruistabel van de ingelezen rapportdata

Maak kruistabel van de ingelezen rapportdata
}
\examples{
\dontrun{
long_format <- lims_report_data(project = c("I-19W001-01"))
XTAB_format <- lims_report_xtab(long_format)
}
\dontrun{
conn <- lims_connect()
reportdata <- read_lims_data(conn, project = c("I-19W001-01"))
xtab <- lims_report_xtab(reportdata)
}
}
