% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_LIMSDB.R
\name{limsdwh_connect}
\alias{limsdwh_connect}
\title{Connect to Data warehouse}
\usage{
limsdwh_connect(server = "inbo-sql08-prd.inbo.be",
  database = "W0003_00_Lims")
}
\arguments{
\item{server}{naam van de server van het LIMS datawarehouse. Staat default correct}

\item{database}{naam van de database van het LIMS datawaerhouse. Staat default correct}
}
\value{
database connection object
}
\description{
Connect to Data warehouse
}
\examples{
\dontrun{
conn <- lims_connect()
sql <- "select top(10) * from factResult"
dfResults <- dbGetQuery(conn, sql)
str(dfResults)
}
}
