% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims_data_limsdwh_report.R
\name{limsdwh_report}
\alias{limsdwh_report}
\title{Haal resultaten uit het LIMS datawarehouse}
\usage{
limsdwh_report(conn, project = NULL, start = NULL, end = Sys.Date(),
  analysis = NULL, format_dec = ",", show_query = FALSE)
}
\arguments{
\item{conn}{connectiestring naar de database. Wordt aangemaakt via dwh_connect() of manueel}

\item{project}{naam van een project, bv "I-19W001-01", of meerdere namen als vector bv. c("I-19W001-01", "I-19W001-02"). Als derde mogelijkheid kan je ook 1 project selecteren met SQL joker karakters als \% en _. Als je alle projecten van contract 19W001 wil kan je dit als volgt specificeren: "\%19W001\%"}

\item{start}{vanaf welke monsternamedatum moet gezocht worden (standaard NULL, is vanaf 2014). De verwachte vorm is yyyy-mm-dd}

\item{end}{tot welke monsternamedatum moet gezocht worden (standaard vandaag). De verwachte vorm is yyyy-mm-dd}

\item{analysis}{naam van een analyse (LimsAnalysisName). Net zoals bij project kan je een vector met meerdere analyses kiezen, of 1 analysenaam doorgeven met SQL jokercharacters \% of _.}

\item{format_dec}{hoe wordt in het geformatteerd resultaat het decimaalteken weergegeven: "," of "."}

\item{show_query}{Moet er een print getoond worden van welke query gebruikt werd om de data te bevragen?}
}
\value{
data.frame uit het lims datawarehouse met alle analyseresultaten die gerapporteerd worden door het labo
}
\description{
Haal resultaten uit het LIMS datawarehouse
}
