% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sql_report_query.R
\name{parse_sql_report_query}
\alias{parse_sql_report_query}
\title{Parse the query based on the template information and the chosen projects}
\usage{
parse_sql_report_query(template, project)
}
\arguments{
\item{template}{dataset containing the template_information.
The structure should be exact as the file shipped with the package}

\item{project}{character string of projects to filter}
}
\value{
SQL server query
}
\description{
Parse the query based on the template information and the chosen projects
}
\examples{
{
temp <- get_report_config_info(template = "default")
query <- parse_sql_report_query(temp, "I-19W001-01")
query
}

}
